Move test code to upstairs lib
Test code had followed the main function for both upstairs
and client instead of staying (where they belong) with the lib file.
Moved them all back, and fixed a few clippy complaints.
